requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.3.2
pdf2image==1.17.0
Pillow==10.2.0
